removed unused imports, moved test into package
